+++ apps/docs/components/doc/preview/codePreview.tsx
@@ -1,8 +1,8 @@
 // import * as hookRaw from "./hook?raw";
 import type { FC, CSSProperties } from "react";
 import { Sandpack, SandpackInternal } from "@codesandbox/sandpack-react";
-// import { githubLight } from "@codesandbox/sandpack-themes";
-import { atomDark } from "@codesandbox/sandpack-themes";
+import { githubLight } from "@codesandbox/sandpack-themes";
+// import { atomDark } from "@codesandbox/sandpack-themes";
 
 export interface Props {
   code: Record<string, string>;
@@ -16,14 +16,14 @@ export const CodePreview: FC<Props> = (props) => {
   return (
     <div className="my-7">
       <Sandpack
-        theme={atomDark}
+        theme={githubLight}
         options={{
           // editorHeight: 600,
           ...props.options,
         }}
         customSetup={{
           dependencies: {
-            "@orderly.network/hooks": "latest",
+            "@orderly.network/hooks": "1.0.4",
 
             axios: "latest",
             ...dependencies,