+++ apps/docs/example/wallet_modal/walletButton.example.tsx
@@ -1,6 +1,6 @@
 import { useWalletConnector } from "./walletConnectProvider";
 
-export const WalletButtonExample = () => {
+export const WalletButton = () => {
   const wallet = useWalletConnector();
   return (
     <button