+++ apps/docs/example/wallet_modal/walletConnectProvider.example.tsx
@@ -8,9 +8,12 @@ import {
 
 import {
   createWeb3Modal,
-  defaultConfig,
+  defaultWagmiConfig,
   useWeb3Modal,
-} from "@web3modal/ethers5/react";
+} from "@web3modal/wagmi/react";
+
+import { WagmiConfig } from "wagmi";
+import { arbitrum, mainnet } from "viem/chains";
 
 import { useAccount } from "@orderly.network/hooks";
 
@@ -21,32 +24,22 @@ interface WalletConnectContextState {
 const WalletConnectContext = createContext<WalletConnectContextState>(
   {} as WalletConnectContextState
 );
-
-// 1. Get projectId
+// 1. Get projectId at https://cloud.walletconnect.com
 const projectId = "YOUR_PROJECT_ID";
 
-// 2. Set chains
-const mainnet = {
-  chainId: 42161,
-  name: "Arbitrum One",
-  currency: "ARB-ETH",
-  explorerUrl: "https://etherscan.io",
-  rpcUrl: "https://rpc.ankr.com/arbitrum",
+// 2. Create wagmiConfig
+const metadata = {
+  name: "Web3Modal",
+  description: "Web3Modal Example",
+  url: "https://web3modal.com",
+  icons: ["https://avatars.githubusercontent.com/u/37784886"],
 };
+
+const chains = [mainnet, arbitrum];
+const wagmiConfig = defaultWagmiConfig({ chains, projectId, metadata });
 
 // 3. Create modal
-const metadata = {
-  name: "My Website",
-  description: "My Website description",
-  url: "https://mywebsite.com",
-  icons: ["https://avatars.mywebsite.com/"],
-};
-
-createWeb3Modal({
-  ethersConfig: defaultConfig({ metadata }),
-  chains: [mainnet],
-  projectId,
-});
+createWeb3Modal({ wagmiConfig, projectId, chains });
 
 export const WalletConnectProvider: FC<PropsWithChildren<{}>> = (props) => {
   const { account } = useAccount();