+++ packages/component/src/page/trading/sections/dataList/position/index.tsx
@@ -84,8 +84,6 @@ export const PositionPane = () => {
       .catch(() => {});
   }, []);
 
-  console.log("isloading", loading);
-
   return (
     <PositionsView
       dataSource={