+++ packages/component/src/toast/toast.stories.tsx
@@ -2,25 +2,10 @@ import type { Meta, StoryObj } from "@storybook/react";
 import React from "react";
 
 import toast, { Toaster } from "react-hot-toast";
-import { OrderlyProvider } from "../provider";
-import { MemoryConfigStore } from "@orderly.network/core";
-import { WooKeyStore } from "../stories/mock/woo.keystore";
 
 const meta: Meta<typeof Toaster> = {
   component: Toaster,
   title: "Base/Toast",
-  decorators: [
-    (Story) => {
-      return (
-        <OrderlyProvider
-          configStore={new MemoryConfigStore()}
-          keyStore={new WooKeyStore("testnet")}
-        >
-          <Story />
-        </OrderlyProvider>
-      );
-    },
-  ],
 };
 
 export default meta;