+++ apps/examples/CHANGELOG.md
@@ -1,5 +1,28 @@
 # examples
 
+## 0.1.73
+
+### Patch Changes
+
+- Updated dependencies
+  - @orderly.network/react@0.1.7
+
+## 0.1.72
+
+### Patch Changes
+
+- Updated dependencies
+  - @orderly.network/react@0.1.6
+
+## 0.1.71
+
+### Patch Changes
+
+- Updated dependencies
+  - @orderly.network/react@0.1.5
+  - @orderly.network/hooks@1.0.4
+  - @orderly.network/core@0.1.3
+
 ## 0.1.70
 
 ### Patch Changes