+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @orderly.network/hooks
 
+## 1.0.4
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/types@0.1.3
+  - @orderly.network/core@0.1.3
+  - @orderly.network/perp@1.0.3
+  - @orderly.network/net@1.0.90
+
 ## 1.0.3
 
 ### Patch Changes