+++ packages/onboard/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @orderly.network/onboard
 
+## 0.0.4
+
+### Patch Changes
+
+- update
+- Updated dependencies
+  - @orderly.network/hooks@1.0.4
+  - @orderly.network/types@0.1.3
+
 ## 0.0.3
 
 ### Patch Changes